+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel-event.ts
@@ -2,7 +2,8 @@
 export enum PanelEventTypes {
   DRAG_START,
   DRAG_OVER,
-  DRAG_END
+  DRAG_END,
+  CLOSE
 }
 
 export interface PanelEvent {