+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel-header/panel-header.ts
@@ -1,4 +1,0 @@
-export interface PanelHeaderOptions {
-  title?: string;
-  cssClass?: string;
-}