+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel.component.ts
@@ -7,13 +7,15 @@ import {
   KeyValueDiffer,
   KeyValueDiffers,
   ChangeDetectorRef,
-  DoCheck
+  DoCheck,
+  OnInit,
+  ViewChild,
+  ElementRef
 } from '@angular/core';
 import {DomSanitizer} from '@angular/platform-browser';
 
 import {Subscriptions} from '../subscriptions';
 import {PanelOptions} from './panel';
-import {PanelHeaderOptions} from './panel-header/panel-header';
 import {PanelEvent, PanelEventTypes} from './panel-event';
 
 @Component({
@@ -21,20 +23,36 @@ import {PanelEvent, PanelEventTypes} from './panel-event';
   templateUrl: './panel.component.html',
   styleUrls: ['./panel.component.css']
 })
-export class PanelComponent extends Subscriptions implements OnChanges, DoCheck {
+export class PanelComponent extends Subscriptions implements OnInit, OnChanges, DoCheck {
+
+  @ViewChild('el') el: ElementRef;
 
   @Input() options: PanelOptions;
   @Output() events: EventEmitter<PanelEvent>;
 
-  panelHeaderOptions: PanelHeaderOptions;
   url: string;
+  title: string;
   width: number;
   height: number;
   left: number;
   top: number;
   isActive: boolean;
+  isClose: boolean;
+  timeOut: any;
+  cssClasses: string[];
 
   private _differ: KeyValueDiffer<string, any> = null;
+  private _isFullScreen: boolean;
+
+
+  get isFullScreen(): boolean {
+    return this._isFullScreen;
+  }
+
+  set isFullScreen(value: boolean) {
+    this._isFullScreen = value;
+    this._calcHeight();
+  }
 
   constructor(private _changeDetector: ChangeDetectorRef,
               private _differs: KeyValueDiffers,
@@ -43,14 +61,10 @@ export class PanelComponent extends Subscriptions implements OnChanges, DoCheck 
     this.events = new EventEmitter();
     this.url = 'http://fiddle.sh';
     this.width = window.innerWidth - 100;
-    this.height = Math.floor(window.innerHeight * .3);
-    this.isActive = true;
+    this.title = 'Panel 1';
     this.left = 0;
     this.top = 0;
-    this.panelHeaderOptions = {
-      title: this.url,
-      cssClass: 'active-panel'
-    };
+    this.cssClasses = ['panel', 'panel-primary'];
   }
 
   ngOnChanges(changes: any): void {
@@ -60,6 +74,11 @@ export class PanelComponent extends Subscriptions implements OnChanges, DoCheck 
         this._differ = this._differs.find(value).create(this._changeDetector);
       }
     }
+  }
+
+  ngOnInit(): void {
+    this._calcHeight();
+    this._calcWidth();
   }
 
   ngDoCheck(): void {
@@ -76,8 +95,69 @@ export class PanelComponent extends Subscriptions implements OnChanges, DoCheck 
     }
   }
 
-  onPanelHeaderEvent(event: PanelEvent) {
-    this.events.emit(event);
+  onWindowResize(event: any) {
+    if (this.timeOut) {
+      window.clearTimeout(this.timeOut);
+    }
+    this.timeOut = window.setTimeout(() => {
+      this._calcHeight();
+      this._calcWidth();
+    }, 64);
+  }
+
+  onDragStart(event: any): void {
+    this.events.emit({
+      type: PanelEventTypes.DRAG_START,
+      data: {
+        event: event,
+        panel: this.options
+      }
+    });
+  }
+
+  onDragEnd(event: any): void {
+    this.events.emit({
+      type: PanelEventTypes.DRAG_END,
+      data: {
+        event: event,
+        panel: this.options
+      }
+    });
+  }
+
+  onDragOver(event: any): void {
+    this.events.emit({
+      type: PanelEventTypes.DRAG_OVER,
+      data: {
+        event: event,
+        panel: this.options
+      }
+    });
+  }
+
+  onEvents(event: any) {
+    switch (event.type) {
+      case 'close':
+        this.events.emit({
+          type: PanelEventTypes.CLOSE,
+          data: {
+            panel: this.options
+          }
+        });
+        break;
+    }
+  }
+
+  private _calcHeight() {
+    if (this.isFullScreen) {
+      this.height = this.el.nativeElement.offsetParent.offsetHeight - (48 * 2) - 16;
+    } else {
+      this.height = Math.floor(window.innerHeight * .3);
+    }
+  }
+
+  private _calcWidth() {
+    this.width = window.innerWidth - Math.floor(window.innerWidth * .11);
   }
 
   private _applyChange(item: any): void {
@@ -98,10 +178,19 @@ export class PanelComponent extends Subscriptions implements OnChanges, DoCheck 
         this.top = this.options.top;
         break;
       case 'title':
-        this.panelHeaderOptions.title = this.options.title;
+        this.title = this.options.title;
         break;
-      case 'isActive':
-        this.isActive = this.options.isActive;
+      case 'active':
+        this.isActive = this.options.active;
+        break;
+      case 'fullScreen':
+        this.isFullScreen = this.options.fullScreen;
+        break;
+      case 'cssClass':
+        this.cssClasses = this.options.cssClass;
+        break;
+      case 'canClose':
+        this.isClose = this.options.canClose;
         break;
     }
   }