+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel.module.ts
@@ -1,12 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PanelComponent } from './panel.component';
-import { PanelHeaderModule } from './panel-header/panel-header.module';
 
 @NgModule({
   imports: [
-    CommonModule,
-    PanelHeaderModule
+    CommonModule
   ],
   declarations: [PanelComponent],
   exports: [