+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel.ts
@@ -1,11 +1,14 @@
 export interface PanelOptions {
-  seq: number,
+  seq: number;
   url: string;
+  id?: number;
   width?: number;
   height?: number;
   top?: number;
   left?: number;
   title?: string;
-  isActive?: boolean;
-
+  active?: boolean;
+  fullScreen?: boolean;
+  canClose?: boolean;
+  cssClass?: string[];
 }