+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/select-list/select-list.ts
@@ -3,7 +3,12 @@ export interface SelectListData {
   /**
    * Unique identifier.
    */
-  id?: string | number;
+  id?: number;
+
+  /**
+   * id using for sorting the list.
+   */
+  order?: number;
 
   /**
    * The text which will be displayed.