+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/store/actions.ts
@@ -1,8 +1,0 @@
-
-// Frame Action Constants
-export const ADD_FRAME = 'ADD_FRAME';
-export const REMOVE_FRAME = 'REMOVE_FRAME';
-export const TOGGLE_ACTIVE = 'TOGGLE_ACTIVE';
-export const SHOW_ACTIVE = 'SHOW_ACTIVE';
-export const SHOW_INACTIVE = 'SHOW_INACTIVE';
-