+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/store/id.ts
@@ -1,4 +1,0 @@
-let start = Math.floor(Math.random() * (5000));
-export const id = () => {
-  return ++start;
-};